Update and delete with confirmation box

+++ src/app/app.component.ts
@@ -2,23 +2,30 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { InfoAddEditComponent } from './info-add-edit/info-add-edit.component';
 import { InfosystemService } from './service/infosystem.service';
+import { SnackBarService } from './service/snack-bar.service';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
+import { ConfirmDialogComponent } from './confirm-dialog/confirm-dialog.component';
+
+
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
 export class AppComponent implements OnInit {
-
+  popoverTitle = 'Popover title';
+  popoverMessage = 'Popover description';
+  confirmClicked = false;
+  cancelClicked = false;
   displayedColumns: string[] = ['id', 'firstName', 'lastName', 'phone', 'email', 'cnic', 'dob', 'gender', 'experience', 'education', 'action'];
   dataSource!: MatTableDataSource<any>;
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
 
-  constructor(private _dialog: MatDialog, private _infoService: InfosystemService) { }
+  constructor(private _dialog: MatDialog, private _infoService: InfosystemService, private _snackBar: SnackBarService) { }
 
   ngOnInit(): void {
 
@@ -28,13 +35,12 @@ export class AppComponent implements OnInit {
     const dialogref = this._dialog.open(InfoAddEditComponent);
     dialogref.afterClosed().subscribe({
 
-      next:(val)=>{
-        if(val){
+      next: (val) => {
+        if (val) {
           this.getInfomration();
         }
       }
     })
-
   }
 
   getInfomration() {
@@ -60,16 +66,35 @@ export class AppComponent implements OnInit {
   }
 
   deleteInfomration(id: number) {
-    this._infoService.deleteInfo(id).subscribe({
+    const dialogRef = this._dialog.open(ConfirmDialogComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      console.log(`Dialog result: ${result}`);
+      if (result) {
+        this._infoService.deleteInfo(id).subscribe({
 
-      next: (resp) => {
-        alert(id + " id deleted successfully");
-        this.getInfomration();
+          next: (resp) => {
+            this._snackBar.openSnackBar(id + " id deleted successfully", 'OK');
+            this.getInfomration();
+          },
+          error: console.log
+        })
+      }
+    });
+  }
 
 
-      },
-      error: console.log
-
+  updateInfomration(data: any) {
+    debugger;
+    const dialogref = this._dialog.open(InfoAddEditComponent, {
+      data: data
+    });
+    dialogref.afterClosed().subscribe({
+      next: (val) => {
+        if (val) {
+          this.getInfomration();
+        }
+      }
     })
+
   }
 }
